Add third level refinement abstractor
This level is fully incremental, unlike previous levels. It's core is generalize-aggr, which takes an aggregate and a new trace, and generalizes the aggregate. Then, the abstract-traces-3 function which is tested alongside the previous refinement levels is simply a fold using generalize-aggr over traces.

It turns out that when you're being fully incremental, you actually need to store a bit more information than easily fits in the shallow expression embedding I've been using so far, so this level introduces a new representation, the aggregate struct, which contains a structure skeleton, plus an explicit variable map which provides equality not just between leaf variables, but between every node in the tree. This makes incrementality possible.