Don't overwrite children on generalization
Basically, before when you had some expression (e.g. (+ (+ 2 3) 4)) and
you generalized one of it's children (+ x 4), then it would overwrite
that childs tea, so that other places that referred to the
expression (e.g. (+ 2 3)) would just get the generalized form (e.g. x),
which is pretty much meaningless. Now we create a new tea for the
generalization.