Fix stack issues with large stems
We used to have an issue where huge (very tall, in terms of tree height)
stems would crash the program when the top node was freed after all the
other nodes had already been disowned everywhere else, because it would
disown it's reference to it's children, causing those children to be
freed, disowning their children, causing them to be freed, etc etc, and
since it would do so through the call stack, calling disownSV to disown,
which would call cleanupStemNode, which would call disownSV, it
overflowed the stack and crashed. Now, we do all the work of both
functions in disownSV, and instead of using recursive calls and the call
stack to cascade frees, we use an explicit queue, defined in
types/hg_queue.c/h. This approach seems to work.